Move all *ParameterDefinition types into own parameters/definition_types.ts file.
