+++ src/manifest/manifest_schema.ts
@@ -1,9 +1,7 @@
 import { ISlackDatastore } from "../datastore/types.ts";
 import { ISlackFunctionDefinition } from "../functions/types.ts";
-import type {
-  ParameterDefinition,
-  ParameterSetDefinition,
-} from "../parameters/types.ts";
+import type { ParameterSetDefinition } from "../parameters/types.ts";
+import type { ParameterDefinition } from "../parameters/definition_types.ts";
 import { OAuth2ProviderTypeValues } from "../schema/providers/oauth2/types.ts";
 import type { ICustomType } from "../types/types.ts";
 import { ISlackWorkflow } from "../workflows/types.ts";
